检查用户是否后台用户

+++ src/main/java/com/acme/acmemall/service/impl/UserService.java
@@ -112,7 +112,7 @@ public class UserService implements IUserService {
      */
     @Override
     public Boolean checkAdmin(long userId) {
-        int count = (int) userDao.findRoleByUserId(userId).get(0).get("cnt");
+        long count = (long) userDao.findRoleByUserId(userId).get(0).get("cnt");
         return count > 0 ? Boolean.TRUE : Boolean.FALSE;
     }
